Move import_export service to backup module
