+++ src/sentry/backup/imports.py
@@ -26,12 +26,7 @@ from sentry.backup.dependencies import (
 )
 from sentry.backup.helpers import Filter, ImportFlags, Printer
 from sentry.backup.scopes import ImportScope
-from sentry.db.models.paranoia import ParanoidModel
-from sentry.models.importchunk import ControlImportChunkReplica
-from sentry.models.orgauthtoken import OrgAuthToken
-from sentry.models.outbox import OutboxCategory, OutboxFlushError, OutboxScope, RegionOutbox
-from sentry.nodestore.django.models import Node
-from sentry.services.hybrid_cloud.import_export.model import (
+from sentry.backup.services.import_export.model import (
     RpcFilter,
     RpcImportError,
     RpcImportErrorKind,
@@ -39,7 +34,12 @@ from sentry.services.hybrid_cloud.import_export.model import (
     RpcImportScope,
     RpcPrimaryKeyMap,
 )
-from sentry.services.hybrid_cloud.import_export.service import ImportExportService
+from sentry.backup.services.import_export.service import ImportExportService
+from sentry.db.models.paranoia import ParanoidModel
+from sentry.models.importchunk import ControlImportChunkReplica
+from sentry.models.orgauthtoken import OrgAuthToken
+from sentry.models.outbox import OutboxCategory, OutboxFlushError, OutboxScope, RegionOutbox
+from sentry.nodestore.django.models import Node
 from sentry.silo.base import SiloMode
 from sentry.silo.safety import unguarded_write
 from sentry.utils.env import is_split_db
